+++ geomancy/checks/__init__.py
@@ -4,13 +4,15 @@ from .base import CheckBase, CheckException
 from .env import CheckEnv
 from .path import CheckPath
 from .exec import CheckExec
+from .platform import CheckPlatform
 from .python import CheckPythonPackage
 
 __all__ = (
-    "CheckBase",
-    "CheckException",
-    "CheckEnv",
-    "CheckPath",
-    "CheckExec",
-    "CheckPythonPackage",
+    CheckBase,
+    CheckException,
+    CheckEnv,
+    CheckPath,
+    CheckExec,
+    CheckPlatform,
+    CheckPythonPackage,
 )